music/create_covers/pyproject.toml
[project]
name = "create-covers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pillow>=11.2.1",
    "requests>=2.32.3",
]
